changed all stripe routes from / to /stripe/ and gave them their own directory to reflect it. Also added a section to readme.md that shows how to jumpstart testing with stripe and the backend
